Adicionando configurações do axios e adicionandoandnasdndo token de acesso a ele
